change variable name (count) to avoid error in GUI
`count()` is a function in MATLAB, but you're also using `count` as a name of a structure array whose elements are accessed with the syntax `count(n).x`.  MATLAB allows overloaded functions - where two functions the same name, but are different - and so in your original code MATLAB doesn't care about your variable `count`.  But, when you put the code into the GUI, MATLAB pre-evaluates the code, sees `count(n).x` and thinks the code is trying to access the function `count()` using dot notation (`.x`). Changing the variable name `count` to `colonyCount` avoids the error.